Make source_code_reference_provided a failure
Prior to this commit, the
`slsa_source_correlated.source_code_reference_provided` policy rule
generated a warning if the expected source code reference was not
provided in the input. When that information is not available, the other
policy rules in the package will pass because there is nothing to match
against. Given that it's easy to ignore errors in different contexts,
this behavior may give users a false sense of security.

This commit changes the policy rule to emit a violation instead of a
warning, making it obvious when an incorrect validation call needs to be
corrected.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>