Merge rule_data.trusted_tasks into trusted_tasks
Ernesto (reasonably IMO) assumed that this would work and I can't
think of a good reason that it shouldn't, especially since it's such
a small patch.

It's true that we're moving towards the "trusted artifacts", aka
"-oci-ta" tasks which mean we can permit custom tasks without
requiring they they are trusted. But still, it's reasonable for
teams to want to ensure their custom tasks are trusted, so I like
the idea of make it more convenient.

Ref: https://issues.redhat.com/browse/EC-709