Add disallowed package external references rule
The purpose of this is to ban packages that have an externalReference in
their cyclonedx sbom that match some regex in the policy rule data.

This is meant to be useful in particular in conjunction with the
[oci-copy](https://github.com/konflux-ci/build-definitions/tree/main/task/oci-copy)
task. That task permits copying arbitrary binary content for the
purposes of distributing it via an OCI registry. It is not safe in
general. Policies should generally forbid it, and not treat it as a
trusted task.

In scenarios where it is deemed permissible, the task should be trusted,
but only if this rule is also in place which will put limits on what
artifacts can be redistributed.

Ref: https://issues.redhat.com/browse/EC-676