Redefine `DOCKERFILE` parameter check
The policy no longer requires that the buildah Task is named "buildah",
but rather looks at the Tasks that build images and selects those that
have the `DOCKERFILE` parameter instead. With that the rule
`buildah_task_has_dockerfile_param` doesn't make much sense any more and
it has been removed.

Resolves: #1027