Add some checks for RHTAP Jenkins attestations
The motivation is to have some real checks in the "@rhtap-jenkins"
collection so I can use them for the RHTAP Jenkins sample templates
development.

I didn't give a lot of thought to the checks themselves - the
invocation id was chosen somewhat arbitrarily - or how we'll
organize things when/if we have many more rules for RHTAP Jenkins
builds, but let's just get something working to start with and
iterate as required.

Ref: https://issues.redhat.com/browse/EC-746