Consider finding counts in addition to status
The status held in the `result` field of the test output is what was
solely considered, so when a test resulted in ERROR
(`"result": "ERROR"`) but in addition to that also had failures, e.g.
`"result": "ERROR", "failures": 3` the failures were not reported.

This changes the logic so that the individual finding counts are
reported as well as the outcome held in `result` field.

Reference: https://issues.redhat.com/browse/EC-620