Don't rely on package annotation in result_helper
I can't remember if there was a good reason for using the package
annotation, but right it seems like we don't need it. The path for
the rule annotation includes the package name, so we can just use
that directly.

Also this avoids the incorrect behavior when the package is not
annotated, (i.e. the bug reported in this Jira).

As mentioned in the commentary, I have some concerns related to the
fact that we're making assumptions about package naming conventions,
i.e. "policy.some_category.some_name". My hunch is there are still
some assumptions we need to stop making, but let's consider that in
future PRs.

Ref: https://issues.redhat.com/browse/EC-831