Remove Red Hat Manifests package
Instead of inspecting the image's filesystem for the presence of certain
files, we'll rely on the expected Tasks to do this. As noted in EC-823,
such tests already exist.

This commit removes the Red Hat Manifests package that is responsible
for ensuring the sbom-purl.json is found within the image.

Ref: EC-823

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>