Check all rpms in sboms have a known repo id
For all components in all sboms that are rpms, require they have a
repository_id value in their purl, and require that the
repository_id value is in in the big list of known repository_ids.

Includes a bit of extra work to limit the number of violations
produced, since I'm expecting there to be either none of many
hundreds.

As mentioned in a Todo in the comments, this is not yet added to the
redhat collection, but it will be added soon in an upcoming PR.

Ref: https://issues.redhat.com/browse/EC-848