Add regal wrapper for VS Code
This commit adds a regal wrapper which includes the custom rego
functions provided by EC CLI. Since the EC CLI does not wrapp regal,
like it does for OPA, e.g. `ec opa`, a simple regal.go is used to create
a version of regal with support for the custom functions.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>