Use PURL's subpath when matching disallowed pkgs
There are certain repositories that defined multiple packages. These
packages may have different attributes, e.g. different licenses. As
such, it is desirable to disallow certain packages from those repos
while still allowing others.

A good example of this is the github.com/hashicorp/consul repo. The
top-level license is quite restritive, but the one under the `/api`
subpath is not.

This commit changes the sbom_cyclonedx.allowed policy rule to take into
account the subpath attribute. The subpaths defined in the items of the
disallowed_packages list must be an exact match to the subpath of
components defined in the SBOM.

To more easily handle the case where all the subpaths must be
disallowed, the special `"*"` subpath value can be used.

Ref: EC-870

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>