Handle missing component `purl` property in SBOM
Seems that we can't rely on all components containing the `purl`
property so when the failure message is generated for the a component
that does not have the `purl` component the rule execution stops and a
violation that should have been reported is not reported.

We can use the `name` property in those cases which is required and
should be present.