Make the task expiry message customizable
Along with the expires-on annotation, Task authors can optionally define
the expiry-message annotation to override the default expiry message.
The default is still "Upgrade to a newer version of the Task."

This required moving the 'solution' part of the rule into the
'failure_msg' instead, because 'solution' does not support templating.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>