Limit sbom rpm repo id checks to cachi2 content
Rather than require all rpm components found in the sbom to have a
present and allowed repo id, we now require only the rpm components
that were created by cachi2.

The motivation: We know that Syft does not create purls with the
repo id value set when it produces sbom components for rpms. We can
and do get syft created rpm components in our sboms, which would
always fail the check if they were considered.

Some day syft might support adding the repo id to the rpm purls it
creates, but at this stage it's only a cachi2 thing.

Ref: https://issues.redhat.com/browse/EC-901