Bundle CycloneDX schemas
The CycloneDX schema references two other schemas. These are fetched
dynamically when the schema is first compiled in rego. This is not ideal
because it requires accessing an external server. This commit takes
those additional schemas and bundles[1] them making it possible to
validate a CycloneDX SBOM completely offline.

[1] https://json-schema.org/understanding-json-schema/structuring#bundling

Fixes #1185

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>