Make sure `ec` is built with network
In #1188 tests were made to run without network. This works if the local
golang module cache already contains the required dependencies, i.e can
be built offline. When a dependency is not available from the cache, due
to the offline nature of the namespace the build fails.

This circumvents that by invoking the equivalent of `ec opa version` so
that the `ec` command is build outside the offlined namespace.