SBOM loading optimization
When we added support for SPDX SBOMs we added the same fallback as we
have done for CycloneDX, i.e. if the SBOM attestation or the
`SBOM_BLOB_URL` don't yield SBOM data, we extract the image under test
and look for the SBOM data in the image's file system.

Given that the SPDX SBOMs are not produced currently, the fallback of
loading from the extracted image file system is always taken, and the
potentially expensive operation of downloading the image layers and
extracting the files is performed.

With this change we don't look for SPDX SBOMs in the image under test at
all, there is no need to add this functionality, we're expecting that
the version of the Task that will produce SPDX SBOMs will do so by
uploading to the registry and emit the `SBOM_BLOB_URL`.

Reference: https://issues.redhat.com/browse/EC-931