Effective on based on leeway
Sets the `effective_on` when vulnerabilities are not reported as
blocking violations due to the leeway policy. Now the evaluation
considers a period instead of the leeway cut-off date, so the
functions/expressions can be reused between the different rules. This
makes the violations resulting from vulnerabilities found within the
leeway period effectively a warning.

The zero period refers to the period with no leeway, i.e. from beginning
of (Unix) time till the current effective date; and the configured
period includes the time up to the leeway cut-off date, i.e. from
beginning of (Unix) time up to the leeway cut-off date.

Reference: https://issues.redhat.com/browse/EC-838