Flatten package names
This commit changes the package name of all the policy rules to a single
level, e.g. `release.cve` becomes `cve`. This better aligns with how
these packages are used today. It moves the EC ecosystem closer to
using package names as they are defined without having to rely on
dynamically alterations to the package name.

All package shortening code is removed from this repository. The package
name used is now an exact match to the value provided to the `package`
keyword in the rego files. (Note that package shortening still exists in
the EC CLI, but that is expected to be removed soon as well.)

The docs generation code is modified to take into account the file path
of the policies in addition to its annotations. This is needed to
distinguish between the policies in different directories.

The "Package full name" section from the docs is removed since it is now
exactly the same as the value of "Package name".

Although policy rules under different directories are never used
together, the tests in this repo load them all at once. As such, the
package names have to be unique across the different directories. The
one case where this caused a conflict was with
`policy/build_task/labels` and `policy/release/labels`. To mitigate
this, `policy/build_task/labels` is renamed to
`policy/build_task/build_labels`. This is a breaking change but given
that this policy is not widely used, this is ok.

Ref: EC-864

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>