Add `lib.json` package with JSON schema validation
Consolidates the repeated logic of JSON schema validation in the
`json.validate_schema` function. Namely, support for arrays --
by marshaling to a string first.

Also categorizes the schema violations by severity (`failure` vs
`warning`).

This is taken advantage of when unexpected additional properties are
encountered in the provided document, which are categorized as a
`warning` instead of `failure`.

Resolves: #1157