Ignore missing SBOM for Image Indexes
Currently, Konflux does not create an SBOM for Image Indexes:
https://issues.redhat.com/browse/KONFLUX-4330

Until then, do not trigger a violation when an SBOM is not found for
such images.

Today, it's not possible to determine if the image being validated is an
Image Index or an Image Manifest, see
https://github.com/enterprise-contract/ec-cli/issues/2121. The Image
Index detection is done via Konflux-specific heuristics as a workaround.

Fixes #1210
Resolves: EC-996

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>