Emit term for base_image_permitted
This commit changes the policy rule
`base_image_registries.base_image_permitted` so it includes a term in
its result. The term is the repository portion of the base image
reference - everything but the tag and digest.

This allows users to allow certain base image registries as an
exception, especially if time bound, e.g.:

```yaml
volatileConfig:
  exclude:
    - value: 'base_image_registries.base_image_permitted:my-shady-registry.io/foo/bar'
      effectiveUntil: '2024-12-01T00:00:00Z'
```

Ref: EC-649

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>