Fix step_images.step_images_accessible check
Although the implementation of ec.oci.image_manifest does return nil if
the Image Manifest is not found, rego converts that to no value at all.
Thus, asserting that the returned value is nil will never be true
causing this policy rule to always pass.

This commit changes the rule evaluation to check for the lack of a value
instead.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>