feat(ec-965): verify a set of pullspecs related to an image
We were previously only able to verify certain pullspecs at build-time
which resulted in a failed test. This would prevent multiple images from
being built and all staged for release. Even if the image references are
valid at release time, the check used was old and the results are not
updated without a rebuild.

If a task has a result named RELATED_IMAGES_DIGEST, all items in the
pushed file will be checked for validity.

Signed-off-by: arewm <arewm@users.noreply.github.com>