Rework CVE policies to allow per CVE exceptions
This commit changes the CVE policy rules to inspect the full CVE report
available in the OCI registry. By doing so, the policies now have the
list of CVEs found, instead of just a summary.

The policy rules have been rewritten to emit the CVE name as the term.
This allows users to add exceptions for a particular CVE. Previously, it
was only possible to add exceptions for a certain CVE level/severity.

The leeway feature was also tweaked. Now, every violation contains an
`effective_on` which is the issue date of the vulnerability plus any
leeway days that may have been configured. The policies rely on the EC
CLI to downgrade violations to warnings when the effective_on date is in
the future.

Fixes #1182
Ref: EC-921

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>