Fix SBOM count condition for base_image_info_found
Previously, the check asserted that there was at least 1 CycloneDX SBOM.
Use the 'sbom.all_sboms' property instead to also count SPDX SBOMs.

Also reword the description of the check. Make it less specific to avoid
having to explain where the base images are in each SBOM format. All the
check does is assert that there are some SBOMs, anyway.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>