Include snapshot digests in trusted build digests
Actually I'm not sure I understand the use case here, but I think
this is what the requirements are.

We have the snapshot available in the input, so it's quite easy to
find the digests of the other snapshot components. Use image.parse
from `policy/lib/image/image.rego` to extract just the image digest.

Ref: https://issues.redhat.com/browse/EC-1086