Use binaries for OPA IDE integration
The OPA VSCode plugin relies on the OPA and Regal binaries to be
available on the system. The exact version of those matter. In fact, the
OPA binary used has to be the one wrapped by the EC CLI, i.e. `ec opa`.
To make this easier, we track those dependencies in go.mod and use bash
scripts to call `go run ...` on the corresponding modules. This works
ok, but it's kind of slow, and it's likely the reason the VSCode plugin
crashes all the time.

This commit changes the bash script to no longer use `go run`. Instead,
it introduces Make targets to create the required binaries. The bash
scripts are modified to automatically call the Make targets, which
should be a no-op if the binaries don't need to be created, and then
invoking the binary directly.

This should give us a snappier development experience.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>