Shorten some long unflat package names
The files being fixed here were using the old convention of nested
namespace package names, which we used to do, but stopped doing
in https://github.com/enterprise-contract/ec-policies/pull/1198 .

The symptom is a little unexpected -- some missing
description/solution text in a violation report due to how the the
metadata lookup is done by the cli. This could actually could be a
separate bug, out of scope in this PR, but possibly in scope for
https://issues.redhat.com/browse/EC-945 which I am also looking
at currently.

Anyway, this change should be enough to fix EC-1233.

Note: The collection.rhtap_multi_ci change might not have any
impact, but I'll change it anyhow so it's consistent with the
others.

Ref: https://issues.redhat.com/browse/EC-1233