Remove effective_on logic from trusted tasks rules
The 'effective_on' field was introduced to deal with sets of required
tasks that would become effective some time in the future, so it was not
meant to handle the validity of the trusted tasks.

For this reason it has been removed from the trusted tasks, and the
rules that relied on that were migrated to the 'expires_on' field.

Ref: https://issues.redhat.com/browse/EC-424