feat: Bump spdx schema to 2.3.1
In the SPDX 2.3 schema there have been couple of breaking changes
introduced that made the schema not compatible with older version. We
found out that a 2.3 schema refuses a valid documents generated by the
spdx-tools libraries due to this incompatibility.

This commit introduces a schema 2.3.1 which is backward compatible and
makes some enums compatible with multiple supported values.

One of the examples fixed in this schema is:

"PACKAGE-MANAGER" vs "PACKAGE_MANAGER" for referenceCategory field.

Related Github issue: https://github.com/spdx/spdx-spec/issues/792
Signed-off-by: Ales Raszka <araszka@redhat.com>