Create a page for every policy package
Create a page template for policy packages pages and generate a page for
each package. The page content is as it is in the policy.template, and
includes package description, package name and the included rules.

The pages are placed in a folder named "packages", and the file naming
convention is ${SCOPE}_${PACKAGE_NAME}.adoc.

Ref: https://issues.redhat.com/browse/EC-984