Refactor collected package annotations for rules
An array of package titles is created so they can be added to the
Rule.Annotations.Custom field.
This refactor changes the array so it stores the whole package
annotation, to make other package information available for later use.

In addition, exclude a rule (and print a message) when a package path
for the rule is missing.