Add missing metadata for new git_branch package
When we generate the docs, we look for package metadata to use in
the documentation for the package.

This should fix the "Error: Package path 'data.git_branch' not found
for rule 'policy/release/git_branch/git_branch.rego:6" message in
the brand new docs generation in PR #1408.

See also PR #1360 where this package was added recently.

Tangentially related to...
Ref: https://issues.redhat.com/browse/EC-984