Skip single image indexes for rpm unique ver check
It's not 100% clear that this is a good idea, but I think it will
help by avoiding a confusing violation in an edge-case situation.

The edge case:
- A single image has multiple versions of the same rpm in its SBOM
  (for some reason or another).
- That image was built as a multi-arch image even though there's
  only one image.

Note that we're not saying that having multiple versions of the same
rpm in the SBOM is typical or normal, (but maybe it can happen due
to multi-stage builds with different base images..?), we're just
saying that we don't want them to produce this particular violation.

The scenario where a multi-arch image with more than one image has
more than one version of the same rpm within a single image is still
going to produce a violation here. If that does become an issue
we'll likely need to revisit this. I can imagine a solution where we
change the logic from "multiple different versions of same rpm
found" to the slightly more specific "multiple different versions of
same rpm found across different arches".

Ref: https://issues.redhat.com/browse/EC-1232