Remove collection metadata from task/results pkg
Discussed in RH internal slack thread here:
https://redhat-internal.slack.com/archives/C031J4KBFME/p1749498678630749

I'm not certain, but I don't think these were added intentionally.

Additionally, the docs rendering, e.g. see
https://conforma.dev/docs/policy/release_policy.html currently
doesn't handle collections well if they're not under policy/release.

We should fix that, but in the short term, this will fix a broken
we've had for a while.

See also #1408 which revealed this problem, and which goes some way
towards fixing it.

Semi-related to...
Ref: https://issues.redhat.com/browse/EC-984