Ignore tag on pre-build sbom ref comparison
We're expecting the value of SCRIPT_RUNNER_IMAGE_REFERENCE to be an
image ref, the policy check is to make sure that image ref appears in the
SBOM. It was doing a direct string comparison, and because the sbom
purl doesn't include the tag, the comparison was failing.

For example, "quay.io/foo/bar:latest@sha256:abc" compared to
"quay.io/foo/bar@sha256:abc" in the SBOM.

After this change, those will be considered equal, and the
`pre_build_script_task.pre_build_script_task_runner_image_in_sbom`
will not be produced.

Note: It would be reasonable to not do this and suggest that the tag
should be removed from the image ref before it gets added to the
SCRIPT_RUNNER_IMAGE_REFERENCE, but I figured let's adhere to the
princple of least surprise and make Conforma handle it the way
someone might expect it to.

See also: https://issues.redhat.com/browse/EC-1242

Ref: https://issues.redhat.com/browse/KFLUXSPRT-3853